requirements.txt
streamlit
google-generativeai
pdfplumber